+++ packages/nextjs/src/config/util.ts
@@ -1,0 +1,29 @@
+import * as fs from 'fs';
+import { sync as resolveSync } from 'resolve';
+
+/**
+ * Returns the version of Next.js installed in the project, or undefined if it cannot be determined.
+ */
+export function getNextjsVersion(): string | undefined {
+  const nextjsPackageJsonPath = resolveNextjsPackageJson();
+  if (nextjsPackageJsonPath) {
+    try {
+      const nextjsPackageJson: { version: string } = JSON.parse(
+        fs.readFileSync(nextjsPackageJsonPath, { encoding: 'utf-8' }),
+      );
+      return nextjsPackageJson.version;
+    } catch {
+      // noop
+    }
+  }
+
+  return undefined;
+}
+
+function resolveNextjsPackageJson(): string | undefined {
+  try {
+    return resolveSync('next/package.json', { basedir: process.cwd() });
+  } catch {
+    return undefined;
+  }
+}
